requirements.txt
attrs>=23.1.0
PyYAML>=6.0
rich>=13.0
pytest>=7.0
